Configure file

+++ src/sun_obj_annotations.py
@@ -1,3 +1,5 @@
+# Get sun annotation data
+
 import os
 import xmltodict
 import json
@@ -6,22 +8,27 @@ config = {}
 with open('config.json', 'r') as fp:
     config = json.load(fp)
 
-root_dir = os.path.join(config['path'], 'data/SUN2012pascalformat')
-img_dir = os.path.join(root_dir, 'JPEGImages')
-ann_dir = os.path.join(root_dir, 'Annotations')
-set_dir = os.path.join(root_dir, 'ImageSets', 'Main')
+if os.path.exists(os.path.join(config['path'], 'data/sun2012_ann.json')):
+    pass
+else:
+    print('Annotations data not found!')
 
-annotations = {}
-for f in os.listdir(ann_dir):
-    try:
-        doc = {}
-        with open(os.path.join(ann_dir, f)) as fs:
-            doc = xmltodict.parse(fs.read())
+    root_dir = os.path.join(config['path'], 'data/SUN2012pascalformat')
+    img_dir = os.path.join(root_dir, 'JPEGImages')
+    ann_dir = os.path.join(root_dir, 'Annotations')
+    set_dir = os.path.join(root_dir, 'ImageSets', 'Main')
 
-        name = doc['annotation']['filename'].replace('.jpg','')
-        annotations[name] = doc
-    except:
-        print(f)
+    annotations = {}
+    for f in os.listdir(ann_dir):
+        try:
+            doc = {}
+            with open(os.path.join(ann_dir, f)) as fs:
+                doc = xmltodict.parse(fs.read())
 
-with open(os.path.join(config['path'], 'data/sun2012_ann.json'), 'w') as fp:
-    json.dump(annotations, fp, sort_keys=True, indent=4)
+            name = doc['annotation']['filename'].replace('.jpg','')
+            annotations[name] = doc
+        except:
+            print(f)
+
+    with open(os.path.join(config['path'], 'data/sun2012_ann.json'), 'w') as fp:
+        json.dump(annotations, fp, sort_keys=True, indent=4)
